Refactors and adds GitHub star button
