+++ bait_biter/_results.py
@@ -2,7 +2,6 @@ import streamlit as st
 
 
 def display_results(video_title: str, video_id: str, implicit_question: str, answer: str) -> None:
-
     col1, col2, col3 = st.columns(3)
     with col1:
         st.empty()
