+++ bait_biter/custom_html/github_button.html
@@ -1,0 +1,3 @@
+<!-- Place this tag where you want the button to render.
+<iframe src="https://ghbtns.com/github-btn.html?user=mennoliefstingh&repo=bait-biter&type=star&count=true&size=large"
+    frameborder="0" scrolling="0" width="170" height="30" title="GitHub"></iframe>
